Contract-Analyzer/wordDocGenerator2/worddocGenerator2/requirements.txt
python-docx>=1.1.0
boto3>=1.34.0
lxml>=5.0.0